BDN-215. Bidding. Changed data types: accuracy is Int, lastfix in sec.

+++ bidon/src/main/java/org/bidon/sdk/config/models/Geo.kt
@@ -12,7 +12,7 @@ data class Geo(
     @field:JsonName("lon")
     val lon: Double?,
     @field:JsonName("accuracy")
-    val accuracy: Float?,
+    val accuracy: Int?,
     @field:JsonName("lastfix")
     val lastFix: Long?,
     @field:JsonName("country")
